+++ src/common/intl/messages.js
@@ -1708,7 +1708,7 @@ const messages = {
 
   // ----- Utils -----
   parseDuration:
-    '{days, plural, =0 {} one {# day} other {# days}} {hours, plural, =0 {} one {# hour} other {# hours}} {minutes, plural, =0 {} one {# minute} other {# minutes}} {seconds, plural, =0 {} one {# second} other {# seconds}}',
+    '{days, plural, =0 {} one {# day } other {# days }}{hours, plural, =0 {} one {# hour } other {# hours }}{minutes, plural, =0 {} one {# minute } other {# minutes }}{seconds, plural, =0 {} one {# second} other {# seconds}}',
 }
 forEach(messages, function (message, id) {
   if (isString(message)) {
