+++ src/xo-app/logs/index.js
@@ -206,7 +206,6 @@ class Log extends BaseComponent {
                       {...returnedValue}
                     />
                   )}
-
                   {id !== undefined && (
                     <span>
                       {' '}
